+++ src/model/Course.java
@@ -19,7 +19,7 @@ public class Course {
 	private String courseName = new String();
 	private String courseStartDate = new String();
 	private String courseEndDate = new String();
-	private int courseGrade;
+	private double courseGrade;
 	//list to hold all assignments related to course
 	private List<Assignment> assignments;
 	
@@ -40,8 +40,8 @@ public class Course {
 	//calculates what percentage is needed to achieve a passed in grade percentage
 	public double calculatePercentNeededForGrade(double gradeDesired) {
 		//get current total points achieved and current overall possible points
-		int currentPoints = getAssignmentActualGradeTotal();
-		int possiblePoints = getAssignmentNeededGradeTotal();
+		double currentPoints = getAssignmentActualGradeTotal();
+		double possiblePoints = getAssignmentNeededGradeTotal();
 		
 		//calculate goal points threshold for course (out of 1000)
 		double overallPointsNeeded = 1000 * (gradeDesired / 100);
@@ -91,14 +91,14 @@ public class Course {
 	}//end getAssignmentByName
 	
 	//sum all assignments grades within the course
-	public int getAssignmentActualGradeTotal() {
+	public double getAssignmentActualGradeTotal() {
 		//exit if there are no assignments yet added to course
 		if(assignments.size() == 0) {
 			return 0;
 		}//end if
 		
 		//used to hold grade values to be averaged
-		int totalGrade = 0;
+		double totalGrade = 0;
 		
 		//loop over all arraylist values
 		for (Assignment assignment :assignments) {
@@ -113,14 +113,14 @@ public class Course {
 	}//end getAssignmentActualGradeTotal
 	
 	//sum all possible assignments grades within the course
-	public int getAssignmentNeededGradeTotal() {
+	public double getAssignmentNeededGradeTotal() {
 		//exit if there are no assignments yet added to course
 		if(assignments.size() == 0) {
 			return 0;
 		}//end if
 		
 		//used to hold grade values to be averaged
-		int totalGrade = 0;
+		double totalGrade = 0;
 		
 		//loop over all arraylist values
 		for (Assignment assignment :assignments) {
@@ -247,11 +247,11 @@ public class Course {
 	}//end setCourseEndDate
 	
 	//For course grade
-	public int getCourseGrade() {
+	public double getCourseGrade() {
 		return courseGrade;
 	}//end getCourseGrade
 	
-	public void setCourseGrade(int grade) {
+	public void setCourseGrade(double grade) {
 		courseGrade = grade;
 	}//end setCourseGrade
